+++ app/smartbrain_3000/web/_app/immutable/chunks/DFuS0s61.js
@@ -1,1 +1,0 @@
-import{J as e}from"./CXkL0Yf_.js";import{n as t}from"./DnnjNT3G.js";var n=e({count:0});async function r(){try{n.count=(await t.unseenScheduleUpdates()).count}catch{}}export{n,r as t};